requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
openai==1.3.7
python-dotenv==1.0.0 